Validate subrequest header further
