+++ packages/next/src/server/lib/server-ipc/utils.ts
@@ -36,37 +36,3 @@ export const filterReqHeaders = (
   }
   return headers as Record<string, undefined | string | string[]>
 }
-
-// These are headers that are only used internally and should
-// not be honored from the external request
-const INTERNAL_HEADERS = [
-  'x-middleware-rewrite',
-  'x-middleware-redirect',
-  'x-middleware-set-cookie',
-  'x-middleware-skip',
-  'x-middleware-override-headers',
-  'x-middleware-next',
-  'x-now-route-matches',
-  'x-matched-path',
-]
-
-export const filterInternalHeaders = (
-  headers: Record<string, undefined | string | string[]>
-) => {
-  for (const header in headers) {
-    if (INTERNAL_HEADERS.includes(header)) {
-      delete headers[header]
-    }
-
-    // If this request didn't origin from this session we filter
-    // out the "x-middleware-subrequest" header so we don't skip
-    // middleware incorrectly
-    if (
-      header === 'x-middleware-subrequest' &&
-      headers['x-middleware-subrequest-id'] !==
-        (globalThis as any)[Symbol.for('@next/middleware-subrequest-id')]
-    ) {
-      delete headers['x-middleware-subrequest']
-    }
-  }
-}
